requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
fonttools==4.60.1
geoopt==0.5.1
gitpython==3.1.45
grpcio==1.76.0
jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
llvmlite==0.45.1
markdown==3.10
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.3
numba==0.62.1
pillow==11.3.0
pydantic==2.12.4
pydantic-core==2.41.5
pynndescent==0.5.13
pyparsing==3.2.5
pytorch-ranger==0.1.1
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentry-sdk==2.44.0
sympy==1.13.1
tensorboard==2.20.0
threadpoolctl==3.6.0
torch-optimizer==0.3.0
triton==3.1.0; platform_system == "Linux"
typing-inspection==0.4.2
umap-learn==0.5.9.post2
wandb==0.23.0
werkzeug==3.1.3
numpy==1.21.1
tqdm==4.46.0
ninja
codecarbon>=3.2,<3.3
omegaconf
hydra-core
pytorch-lightning
torch
geomloss